requirements.txt
pandas==1.0.3
beautifulsoup4==4.9.0
python-dotenv==0.13.0
requests==2.23.0